feat: 상세이미지 끝 (#226)
