+++ frontend/src/components/Common/Modal/ApplyDetailImgsModal/index.jsx
@@ -1,9 +1,31 @@
-import React from "react";
+import React, { useState, useRef, useEffect } from "react";
 import Portal from "../Portal";
-import ApplyImgsSlide from "./ApplyImgsSlide";
+
+import Presenter from './presenter'
 import * as S from "./style";
 
 export default function ApplyImgsModal({ setModalSwitch, currentImg, files}) {
+  const [currentSlide, setCurrentSlide] = useState(currentImg);
+  const slideRef = useRef(null);
+  const nextSlide = () => {
+    if (currentSlide+1 < files.length) { 
+      setCurrentSlide(currentSlide + 1);
+    } else {
+      return;
+    }
+  };
+  const prevSlide = () => {
+    if (currentSlide === 0) {
+      return;
+    } else {
+      setCurrentSlide(currentSlide - 1);
+    }
+  };
+
+  useEffect(() => {
+    slideRef.current.style.transition = "all 0.5s ease-in-out";
+    slideRef.current.style.transform = `translateX(-${currentSlide}00%)`;
+  }, [currentSlide]);
   const closeModal = () => {
     setModalSwitch(false);
 
@@ -13,11 +35,16 @@ export default function ApplyImgsModal({ setModalSwitch, currentImg, files}) {
   return (
     <Portal>
       <S.ModalBackground>
-        <div>
-          <S.ModalBtn onClick={closeModal}>X 닫기</S.ModalBtn>
-          <S.ModalBtn>다운로드</S.ModalBtn>
-        </div>
-        <ApplyImgsSlide currentImg={currentImg} files={files}/>
+        <Presenter
+        closeModal={closeModal}
+        currentSlide={currentSlide}
+        slideRef={slideRef}
+         setModalSwitch={setModalSwitch}
+          currentImg={currentImg}
+           files={files}
+           prevSlide={prevSlide}
+           nextSlide={nextSlide}
+        />
       </S.ModalBackground>
     </Portal>
   );
